backend/requirements.txt
Django==4.2.6
django-cors-headers==4.3.1
django-ninja==1.1.0
